Add basic support for image upload
This still has a long way to go, but the basics are there now.
We can upload an image and access it in the editor via <img> tags
or the basic Markdown image syntax. Still gotta figure out a good way to
allow the user to insert the image urls...